refactor: rename documents file to upload.ts
